Unlocked User Account
Replace "username" with the actual username of the locked account you want to unlock, then run the script in a PowerShell console. If the account is found and successfully unlocked, the script will display a message indicating that the account has been unlocked. If the account is not found, the script will display a message indicating that the account was not found.